+++ functions/src/modules/change-log/changeLog.trigger.ts
@@ -1,0 +1,12 @@
+import {
+  onDocumentWrittenWithAuthContext,
+} from "firebase-functions/v2/firestore";
+import {handleDocumentChange} from "./changeLog.service";
+
+export const logDocumentChanges = onDocumentWrittenWithAuthContext(
+  {
+    document: "{documentPath=**}",
+    region: "us-central1",
+  },
+  handleDocumentChange,
+);